refactored big part of the locationFragmentViewModel by extracting the methods and put them into smaller classes
