+++ app/src/main/java/com/example/concordia_campus_guide/Fragments/LocationFragment/LocationFragmentViewModel.java
@@ -17,8 +17,11 @@ import com.example.concordia_campus_guide.Adapters.DirectionWrapper;
 import com.example.concordia_campus_guide.ClassConstants;
 import com.example.concordia_campus_guide.Database.AppDatabase;
 import com.example.concordia_campus_guide.Global.ApplicationState;
+import com.example.concordia_campus_guide.Helper.CurrentLocation;
 import com.example.concordia_campus_guide.Helper.DrawPolygons;
 import com.example.concordia_campus_guide.Helper.FloorPlan;
+import com.example.concordia_campus_guide.Helper.ManipulateWalkingPoints;
+import com.example.concordia_campus_guide.Helper.POIIcon;
 import com.example.concordia_campus_guide.Helper.PathFinder;
 import com.example.concordia_campus_guide.Models.Building;
 import com.example.concordia_campus_guide.Models.Coordinates;
@@ -61,19 +64,12 @@ public class LocationFragmentViewModel extends ViewModel {
     private Map<String, Building> buildings = new HashMap<>();
     private AppDatabase appDatabase;
     private MutableLiveData<PriorityQueue<WalkingPoint>> poiList = new MutableLiveData<>();
-    private MutableLiveData<List<RoomModel>> roomList = new MutableLiveData<>();
-    private BitmapDescriptor currentPOIIcon;
-    private Location currentLocation;
-
     private FloorPlan floorPlan;
     private DrawPolygons drawPolygons;
 
-    /**
-     * @return return the map style
-     */
-    public int getMapStyle() {
-        return R.raw.mapstyle_retro;
-    }
+    private CurrentLocation currentLocation;
+    private POIIcon poiIcon;
+    private ManipulateWalkingPoints manipulateWalkingPoints;
 
     //EV centerCoordinates
     private LatLng initialZoomLocation = ClassConstants.initialZoomLocation;
@@ -81,10 +77,12 @@ public class LocationFragmentViewModel extends ViewModel {
     public static final Logger LOGGER = Logger.getLogger("LocationFragmentViewModel");
 
     private HashMap<String, List<WalkingPoint>> walkingPointsMap = new HashMap<>();
-    private List<WalkingPoint> walkingPoints;
 
     public LocationFragmentViewModel(AppDatabase appDb) {
         this.appDatabase = appDb;
+        this.poiIcon = new POIIcon();
+        this.floorPlan = new FloorPlan(appDatabase);
+        this.manipulateWalkingPoints = new ManipulateWalkingPoints();
     }
 
     /**
@@ -96,31 +94,8 @@ public class LocationFragmentViewModel extends ViewModel {
      */
     public GeoJsonLayer loadPolygons(GoogleMap map, Context applicationContext) {
         drawPolygons = new DrawPolygons();
-        GeoJsonLayer layer = initLayer(map, applicationContext);
-        drawPolygons.loadPolygons(layer,map,applicationContext,buildings);
-        return layer;
+        return drawPolygons.loadPolygons(map,applicationContext,buildings);
     }
-
-    /**
-     * The purpose of this method is to initiate the layer
-     *
-     * @param map                is the map used in our application.
-     * @param applicationContext is the Context of the LocationFragmentView page
-     * @return the initiated layer or it will throw an exception if it didn't find the
-     * GeoJson File
-     */
-    private GeoJsonLayer initLayer(GoogleMap map, Context applicationContext) {
-        GeoJsonLayer layer = null;
-
-        try {
-            JSONObject geoJsonLayer = ApplicationState.getInstance(applicationContext).getBuildings().getGeoJson();
-            layer = new GeoJsonLayer(map, geoJsonLayer);
-        } catch (Exception e) {
-            LOGGER.log(Level.WARNING, e.getMessage(), e);
-        }
-        return layer;
-    }
-
 
     public Building getBuildingFromGeoJsonFeature(GeoJsonFeature feature) {
         drawPolygons = new DrawPolygons();
@@ -132,7 +107,6 @@ public class LocationFragmentViewModel extends ViewModel {
      * @return Int of drawable resource's bitmap representation
      */
     public void setFloorPlan(GroundOverlay groundOverlay, String buildingCode, String floor, GoogleMap mMap) {
-        floorPlan = new FloorPlan(appDatabase);
         floorPlan.setFloorPlan(groundOverlay,buildingCode,floor,mMap,walkingPointsMap);
     }
 
@@ -167,122 +141,50 @@ public class LocationFragmentViewModel extends ViewModel {
 
     public void setListOfPOI(@PoiType String poiType, Context context) {
         List<WalkingPoint> allPOI = appDatabase.walkingPointDao().getAllPointsForPointType(poiType);
-        setCurrentPOIIcon(poiType, context);
-        this.poiList.postValue(getPOIinOrder(allPOI));
+        poiIcon.setCurrentPOIIcon(poiType, context);
+        this.poiList.postValue(poiIcon.getPOIinOrder(appDatabase,allPOI,currentLocation));
     }
 
-    private PriorityQueue<WalkingPoint> getPOIinOrder(List<WalkingPoint> allPOI) {
-        PriorityQueue<WalkingPoint> orderedList = new PriorityQueue<>((WalkingPoint p1, WalkingPoint p2) -> {
-            Coordinates currentCoordinates;
 
-            if (currentLocation == null) {
-                //This building has inverted lat/lng in order to us th geojsons.
-                Building hallBuilding = appDatabase.buildingDao().getBuildingByBuildingCode("H");
-                Double currentLat = hallBuilding != null ? hallBuilding.getCenterCoordinates().getLatitude() : 0;
-                Double currentLng = hallBuilding != null ? hallBuilding.getCenterCoordinates().getLongitude() : 0;
-
-                //Current location should be inversed: lat->lng and lng->lat
-                currentCoordinates = new Coordinates(currentLng, currentLat);
-            } else {
-                currentCoordinates = new Coordinates(currentLocation.getLatitude(), currentLocation.getLongitude());
-            }
-
-            double distanceFromP1 = p1.getCoordinate().getEuclideanDistanceFrom(currentCoordinates);
-            double distanceFromP2 = p2.getCoordinate().getEuclideanDistanceFrom(currentCoordinates);
-
-            //Compare walking points: If p1 is closer to the current location than p2, it will have a higher position in priority queue
-            if (distanceFromP1 < distanceFromP2) return -1;
-            else if (distanceFromP1 > distanceFromP2) return 1;
-            return 0;
-        });
-        orderedList.addAll(allPOI);
-        return orderedList;
+    public BitmapDescriptor getCustomSizedIcon(String filename, Context context, int height, int width) {
+        return poiIcon.getCustomSizedIcon(filename,context,height,width);
     }
 
+    public BitmapDescriptor getCurrentPOIIcon(){
+        return poiIcon.getCurrentPOIIcon();
+    }
 
     LiveData<List<RoomModel>> getListOfRoom() {
-        return floorPlan.getListOfRoom();
+        return floorPlan.getRoomsInAFloor().getListOfRoom();
     }
 
     LiveData<PriorityQueue<WalkingPoint>> getListOfPOI() {
         return poiList;
     }
 
-    private void setCurrentPOIIcon(@PoiType String poiType, Context context) {
-        currentPOIIcon = getCustomSizedIcon("point_of_interest_icons/poi_" + poiType.toLowerCase() + ".png", context, 60, 60);
-    }
-
-    BitmapDescriptor getCurrentPOIIcon() {
-        return currentPOIIcon;
-    }
-
-
-    public BitmapDescriptor getCustomSizedIcon(String filename, Context context, int height, int width) {
-        InputStream deckFile = null;
-        BitmapDescriptor smallMarkerIcon = null;
-        try {
-            deckFile = context.getAssets().open(filename);
-            Bitmap b = BitmapFactory.decodeStream(deckFile);
-            Bitmap smallMarker = Bitmap.createScaledBitmap(b, width, height, false);
-            smallMarkerIcon = BitmapDescriptorFactory.fromBitmap(smallMarker);
-            deckFile.close();
-        } catch (IOException e) {
-            LOGGER.log(Level.WARNING, e.getMessage(), e);
-        }
-        return smallMarkerIcon;
-    }
-
-    public void drawOutdoorPath(List<DirectionWrapper> outdoorDirections, GoogleMap map) {
-        for (DirectionWrapper directionWrapper : outdoorDirections) {
-            int color = 0;
-            if (directionWrapper.getTransitDetails() != null) {
-                color = Color.parseColor(directionWrapper.getTransitDetails().line.color);
-            }
-            PolylineOptions polylineOptions = stylePolyLine(directionWrapper.getDirection().getTransportType(), color);
-            List<com.example.concordia_campus_guide.GoogleMapsServicesTools.GoogleMapsServicesModels.LatLng> polyline = directionWrapper.getPolyline().decodePath();
-
-            for (int i = 0; i < polyline.size(); i++) {
-                polylineOptions.add(new LatLng(polyline.get(i).lat, polyline.get(i).lng));
-            }
-            map.addPolyline(polylineOptions);
-        }
-    }
-
-    public PolylineOptions stylePolyLine(String type, int color) {
-        PolylineOptions polylineOptions = new PolylineOptions().width(20);
-        if (type.equals(ClassConstants.WALKING)) {
-            polylineOptions.pattern(ClassConstants.WALK_PATTERN);
-        }
-        if (color != 0) {
-            polylineOptions.color(color);
-        } else {
-            polylineOptions.color(Color.rgb(147, 35, 57));
-        }
-        return polylineOptions;
-    }
-
     public RoomModel getRoomByRoomCodeAndFloorCode(String roomCode, String floorCode){
         return appDatabase.roomDao().getRoomByRoomCodeAndFloorCode(roomCode, floorCode);
     }
 
-    public List<WalkingPoint> getWalkingPointsList() {
-        return walkingPoints;
+    public void parseWalkingPointList(AppDatabase appDatabase, RoomModel from, RoomModel to){
+        manipulateWalkingPoints.parseWalkingPointList(appDatabase, from, to, walkingPointsMap);
     }
 
-    public void parseWalkingPointList(AppDatabase appDatabase, RoomModel from, RoomModel to) {
-        PathFinder pf = new PathFinder(appDatabase, from, to);
-        walkingPoints = pf.getPathToDestination();
-
-        List<WalkingPoint> floorWalkingPointList;
-        for (WalkingPoint wp : walkingPoints) {
-            floorWalkingPointList = walkingPointsMap.getOrDefault(wp.getFloorCode(), new ArrayList<WalkingPoint>());
-            floorWalkingPointList.add(wp);
-            walkingPointsMap.put(wp.getFloorCode(), floorWalkingPointList);
-        }
+    public List<WalkingPoint> getWalkingPointsList(){
+       return manipulateWalkingPoints.getWalkingPointsList();
     }
+
+
 
     public void drawShuttlePath(GoogleMap mMap, String polyline) {
         List<LatLng> route = PolyUtil.decode(polyline);
         mMap.addPolyline(new PolylineOptions().addAll(route).color(R.color.colorAppTheme).width(20));
+    }
+
+    /**
+     * @return return the map style
+     */
+    public int getMapStyle() {
+        return R.raw.mapstyle_retro;
     }
 }
