+++ app/src/main/java/com/example/concordia_campus_guide/Helper/FloorPlan.java
@@ -4,8 +4,6 @@ import android.graphics.Color;
 
 import com.example.concordia_campus_guide.ClassConstants;
 import com.example.concordia_campus_guide.Database.AppDatabase;
-import com.example.concordia_campus_guide.Models.Floor;
-import com.example.concordia_campus_guide.Models.RoomModel;
 import com.example.concordia_campus_guide.Models.WalkingPoint;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.model.BitmapDescriptorFactory;
@@ -14,23 +12,30 @@ import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Polyline;
 import com.google.android.gms.maps.model.PolylineOptions;
 import com.google.maps.android.geojson.GeoJsonLayer;
-
 import java.util.HashMap;
 import java.util.List;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
 
 public class FloorPlan {
     private GeoJsonLayer floorLayer;
     private Polyline currentlyDisplayedLine;
-    private MutableLiveData<List<RoomModel>> roomList = new MutableLiveData<>();
-    private HashMap<String, List<WalkingPoint>> walkingPointsMap = new HashMap<>();
     private AppDatabase appDatabase;
+
+    public RoomsInAFloor getRoomsInAFloor() {
+        return roomsInAFloor;
+    }
+
+    public void setRoomsInAFloor(RoomsInAFloor roomsInAFloor) {
+        this.roomsInAFloor = roomsInAFloor;
+    }
+
+    private RoomsInAFloor roomsInAFloor;
 
     public FloorPlan(AppDatabase appDB){
         this.appDatabase = appDB;
+        roomsInAFloor = new RoomsInAFloor();
     }
+
+
 
     /**
      * @param buildingCode it represents which building we will be covering
@@ -43,15 +48,18 @@ public class FloorPlan {
     }
 
     public void setFloorMarkers(String buildingCode, String floor, GoogleMap mMap, HashMap<String, List<WalkingPoint>> walkingPointsMap) {
+
         if (floorLayer != null) {
             floorLayer.removeLayerFromMap();
         }
-        setListOfRooms(buildingCode+"-"+floor);
+
+        roomsInAFloor.setListOfRooms(buildingCode+"-"+floor, appDatabase);
         if (currentlyDisplayedLine != null) {
             currentlyDisplayedLine.remove();
         }
         PolylineOptions displayedPolylineOption = getFloorPolylines(buildingCode + "-" + floor, walkingPointsMap);
         currentlyDisplayedLine = mMap.addPolyline(displayedPolylineOption);
+        String x = "asd";
     }
 
     public PolylineOptions getFloorPolylines(String floorCode, HashMap<String, List<WalkingPoint>> walkingPointsMap) {
@@ -73,14 +81,7 @@ public class FloorPlan {
                 .visible(true);
     }
 
-    public void setListOfRooms(String floorCode) {
-        List<RoomModel> allRoomsOnFloor = appDatabase.roomDao().getAllRoomsByFloorCode(floorCode);
-        this.roomList.postValue(allRoomsOnFloor);
-    }
 
-    public LiveData<List<RoomModel>> getListOfRoom() {
-        return roomList;
-    }
 
 
 }
