--- app/src/main/java/com/example/concordia_campus_guide/Helper/MakersManipulation.java
@@ -1,6 +1,0 @@
-package com.example.concordia_campus_guide.Helper;
-
-public class MakersManipulation {
-
-
-}
