+++ app/src/main/java/com/example/concordia_campus_guide/Helper/MapSetup.java
@@ -1,6 +1,7 @@
 //package com.example.concordia_campus_guide.Helper;
 //
 //import android.annotation.SuppressLint;
+//import android.content.Context;
 //import android.content.res.Resources;
 //import android.util.Log;
 //
@@ -9,17 +10,51 @@
 //import com.example.concordia_campus_guide.Models.Building;
 //import com.example.concordia_campus_guide.R;
 //import com.google.android.gms.maps.GoogleMap;
+//import com.google.android.gms.maps.MapView;
+//import com.google.android.gms.maps.MapsInitializer;
 //import com.google.android.gms.maps.model.MapStyleOptions;
 //
+//import java.util.Map;
+//
+//
 //public class MapSetup {
+//
+//    private GoogleMap mMap;
+//    private Context context;
+//    MapView mMapView;
+//
+//    public MapSetup(Context context){
+//
+//        this.context = context;
+//    }
+//
+//
+//    /**
+//     * The purpose of this method is to init the map and fill it up with the required
+//     * information to display them to the user
+//     */
+//    public void initMap() {
+//        try {
+//            MapsInitializer.initialize(context.getApplicationContext());
+//        } catch (Exception e) {
+//            Log.e(ClassConstants.LOCATION_FRAGMENT_TAG, e.getMessage());
+//        }
+//        mMapView.getMapAsync(googleMap -> {
+//            setMapStyle(googleMap);
+//            mMap = googleMap;
+//            setupPolygons(mMap);
+//            initFloorPlans();
+//            uiSettingsForMap(mMap);
+//            setFirstLocationToDisplay();
+//        });
+//    }
 //
 //    /**
 //     * set up related to UI for the map
 //     *
-//     * @param mMap
 //     */
 //    @SuppressLint("MissingPermission")
-//    private void uiSettingsForMap(GoogleMap mMap) {
+//    private void uiSettingsForMap() {
 //        if (ClassConstants.MY_LOCATION_PERMISSION_GRANTED) {
 //            mMap.setMyLocationEnabled(true);
 //        }
@@ -31,11 +66,11 @@
 //    }
 //
 //
-//    private void setMapStyle(GoogleMap googleMap) {
+//    private void setMapStyle() {
 //        try {
-//            googleMap.setMapStyle(
+//            mMap.setMapStyle(
 //                    MapStyleOptions.loadRawResourceStyle(
-//                            getContext(), getMapStyle()));
+//                            context, getMapStyle()));
 //        } catch (Resources.NotFoundException e) {
 //            Log.e("MAPACTIVITY", "Can't find style. Error: ", e);
 //        }
@@ -48,62 +83,20 @@
 //        return R.raw.mapstyle_retro;
 //    }
 //
-//
 //    /**
-//     * The purpose of this method is display the polygon on the map and
-//     * call the right method for onClick polygon or on click the marker.
-//     *
-//     * @param map is the map to be used in our application
+//     * The purpose of this method is to figure the style of the map to display
 //     */
-//    private void setupPolygons(GoogleMap map) {
-//        mLayer = mViewModel.loadPolygons(map, getContext());
-//        mLayer.addLayerToMap();
+//    private void initFloorPlans() {
+//        Map<String, Building> buildings = mViewModel.getBuildings();
+//        for (Map.Entry<String, Building> entry : buildings.entrySet()) {
+//            String key = entry.getKey();
+//            Building value = entry.getValue();
 //
-//        setupPolygonClickListener();
-//        setupBuildingMarkerClickListener(map);
-//        setupZoomListener(map);
-//        classRoomCoordinateTool(map);
+//            if (value.getGroundOverlayOption() != null)
+//                buildingsGroundOverlays.put(key, mMap.addGroundOverlay(buildings.get(key).getGroundOverlayOption()));
+//        }
 //    }
 //
-//    private void setupZoomListener(final GoogleMap map) {
-//        map.setOnCameraMoveListener(() -> {
-//            if (map.getCameraPosition().zoom > 19) {
-//                mLayer.removeLayerFromMap();
-//            } else {
-//                mLayer.addLayerToMap();
-//            }
-//        });
-//    }
-//    private void classRoomCoordinateTool(GoogleMap map) {
-//        map.setOnMapClickListener(latLng -> Log.i(ClassConstants.LOCATION_FRAGMENT_TAG, "\"coordinates\" : [" + latLng.longitude + ", " + latLng.latitude + "]"));
-//    }
-//
-//
-//
-//    /**
-//     * The purpose of this method is handle the onclick marker
-//     * and to open the info card according to the clicked building.
-//     */
-//    public boolean setupBuildingMarkerClickListener(GoogleMap map) {
-//        map.setOnMarkerClickListener(marker -> {
-//            if (marker.getTag() != null) {
-//                if (marker.getTag() != null && (marker.getTag().toString().contains(ClassConstants.ROOM_TAG) || marker.getTag().toString().contains(ClassConstants.POI_TAG))) {
-//                    String roomCode = marker.getTag().toString().split("_")[1];
-//                    String floorCode = marker.getTag().toString().split("_")[2];
-//                    onRoomClick(roomCode, floorCode);
-//                }
-//                else {
-//                    Building building = mViewModel.getBuildingFromCode(marker.getTag().toString());
-//                    String buildingCode = (marker.getTag()).toString();
-//                    if (getActivity() instanceof MainActivity)
-//                        ((MainActivity) getActivity()).showInfoCard(buildingCode);
-//                    onBuildingClick(building);
-//                }
-//            }
-//            return false;
-//        });
-//        return true;
-//    }
 //
 //
 //
